backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
httpx==0.25.1
pygithub==2.1.1
gitpython==3.1.40
aiofiles==23.2.1
python-multipart==0.0.6
# openai==1.3.0  # Optional - only needed if using OpenAI (demo uses hardcoded responses)
